requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.2
gspread==6.1.2
google-auth==2.34.0
google-api-python-client==2.108.0
python-dotenv==1.0.1
pytz==2023.3
APScheduler==3.10.4
